treating as symlink in post

+++ SPECS/acs-cb-tcltk.spec
@@ -10,7 +10,7 @@ BuildRequires:	ksh, libX11-devel, gcc, make, tar
 #Requires:	expect tk iwidgets tclx tcllib blt tktable
 Requires:	telnet, libX11
 AutoReq:	no
-Conflicts:	tcl
+
 %description
 Tcl, mini SQL, tklib, tclCheck, tkman, tkimg,  itcl, rman, snack, tkcon
 
@@ -37,7 +37,6 @@ sh buildTcltk
 unlink %{_builddir}/alma
 
 %install
-mkdir -p %{buildroot}%{_usr}/lib/tcl8.5/
 mkdir -p %{buildroot}/%{_usr}/local/share/tcltk
 cd %{_builddir}/home/almamgr/%{name}-%{version}/tcltk/
 find lib/ -name *.so | xargs chmod 755 $1
@@ -55,7 +54,12 @@ echo "export PATH" >> %{buildroot}%{_sysconfdir}/profile.d/tcltk-acs.sh
 # /include to usr/local/include. Not yet, as no packages use tcltk headers to compile
 ln -s %{_usr}/local/share/tcltk/bin/tcl %{_usr}/local/bin/seqSh
 ln -s %{_usr}/bin/wish %{_usr}/local/bin/seqWish
-ln -s %{_usr}/local/share/tcltk/lib/tcl8.5/init.tcl %{_usr}/lib/tcl8.5/init.tcl 
+# If folder doesnt exist, create it
+mkdir -p %{_usr}/lib/tcl8.5/
+# If file doesnt exist, create symlink
+if [ ! -f %{_usr}/lib/tcl8.5/init.tcl ]; then
+  ln -s %{_usr}/local/share/tcltk/lib/tcl8.5/init.tcl %{_usr}/lib/tcl8.5/init.tcl 
+fi
 
 
 %preun
@@ -64,11 +68,13 @@ export PATH=$( echo $PATH | sed 's/\%{_usr}\/local\/share\/tcltk\/bin\///g' )
 unset TCLTK_ROOT
 unlink /usr/local/bin/seqSh
 unlink /usr/local/bin/seqWish
-unlink %{_usr}/lib/tcl8.5/init.tcl
+
+if [ -f %{_usr}/lib/tcl8.5/init.tcl ]; then
+  unlink %{_usr}/lib/tcl8.5/init.tcl
+fi
 
 %files
 %attr(755,-,-) %{_usr}/local/share/tcltk/
-%{_usr}/lib/tcl8.5/
 %{_sysconfdir}/profile.d/tcltk-acs.sh
 
 %changelog
